Make control page, and fix index

+++ WheaterSwiftIOS/Model/Wind.swift
@@ -10,5 +10,5 @@ import Foundation
 
 struct Wind: Decodable {
     let speed: Double
-    let deg: Int
+    let deg: Double
 }
